requirements.txt
streamlit
numpy
opencv-python
Pillow
hugchat
ultralytics